refactor(Tooltip): move v1/v2 to src
