+++ packages/components/rollup.config.mjs
@@ -7,8 +7,8 @@ export default rollupConfig({
     actionsV1: './src/__actions__/v1.ts',
     actionsV2: './src/__actions__/v2.ts',
     actionsV3: './src/__actions__/v3.ts',
-    overlaysV1: './src/__overlays__/v1.ts',
-    overlaysV2: './src/__overlays__/v2.ts',
+    overlaysV1: './src/v1-overlays.ts',
+    overlaysV2: './src/v2-overlays.ts',
     overlaysV3: './src/__overlays__/v3.ts',
     utilitiesV3: './src/__utilities__/v3.ts',
     reactAriaV3: './src/__react-aria__/index.ts',
