+++ packages/components/src/Filter/FilterButton/FilterButtonRemovable/FilterButtonRemovable.tsx
@@ -3,8 +3,8 @@ import { useIntl } from '@cultureamp/i18n-react-intl'
 import classnames from 'classnames'
 import { ButtonGroup, ButtonGroupProps } from '~components/ButtonGroup'
 import { FilterTriggerRef } from '~components/Filter/Filter'
+import { Tooltip } from '~components/Tooltip'
 import { Icon } from '~components/__rc__/Icon'
-import { Tooltip } from '~components/__overlays__/Tooltip/v1'
 import { DataAttributes } from '~components/types/DataAttributes'
 import { isRefObject } from '~components/utils/isRefObject'
 import { FilterButton, FilterButtonProps } from '../FilterButton'
