+++ packages/components/src/GuidanceBlock/GuidanceBlock.tsx
@@ -4,9 +4,9 @@ import Media from 'react-media'
 import { Heading, HeadingProps } from '~components/Heading'
 import { SceneProps, SpotProps } from '~components/Illustration'
 import { Text } from '~components/Text'
+import { Tooltip, TooltipProps } from '~components/Tooltip'
 import { Button, ButtonProps } from '~components/__actions__/v2'
 import { Icon } from '~components/__rc__/Icon'
-import { Tooltip, TooltipProps } from '~components/__overlays__/Tooltip/v1'
 import { VariantType } from './types'
 import styles from './GuidanceBlock.module.css'
