+++ packages/components/src/RichTextEditor/RichTextEditor/subcomponents/ToggleIconButton/ToggleIconButton.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import classnames from 'classnames'
-import { Tooltip } from '~components/__overlays__/Tooltip/v1'
+import { Tooltip } from '~components/Tooltip'
 import { OverrideClassName } from '~components/types/OverrideClassName'
 import { ToggleIconButtonMoods } from './types'
 import styles from './ToggleIconButton.module.scss'
