+++ packages/components/src/Table/Table.tsx
@@ -2,8 +2,8 @@ import React, { cloneElement, HTMLAttributes, ReactElement } from 'react'
 import classnames from 'classnames'
 import { Checkbox, CheckedStatus } from '~components/Checkbox'
 import { Heading } from '~components/Heading'
+import { Tooltip } from '~components/Tooltip'
 import { Icon } from '~components/__rc__/Icon'
-import { Tooltip } from '~components/__overlays__/Tooltip/v1'
 import { OverrideClassName } from '~components/types/OverrideClassName'
 import styles from './Table.module.scss'
