--- packages/components/src/__overlays__/Tooltip/v2/index.ts
@@ -1,1 +1,0 @@
-export * from '../v1/Tooltip'
