+++ packages/components/src/__overlays__/Tooltip/v3/_docs/Tooltip.docs.stories.tsx
@@ -6,8 +6,8 @@ import { Input } from '~components/Input'
 import { Label } from '~components/Label'
 import { Text } from '~components/Text'
 import { Button, IconButton } from '~components/__actions__/v2'
-import { Icon } from '~components/__rc__/Icon'
 import { Focusable } from '~components/__overlays__/v3'
+import { Icon } from '~components/__rc__/Icon'
 import { Tooltip, TooltipTrigger } from '../index'
 import * as TestStories from './Tooltip.spec.stories'
