--- packages/components/src/__overlays__/v1.ts
@@ -1,1 +1,0 @@
-export * from './Tooltip/v1'
