+++ packages/components/src/v1-overlays.ts
@@ -1,0 +1,1 @@
+export * from './Tooltip'
