+++ packages/components/src/v2-overlays.ts
@@ -1,0 +1,1 @@
+export * from './Tooltip'
